requirements.txt
Django
numpy
djangorestframework
keras
tensorflow
imutils
sklearn
opencv-python
